pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unitree-sdk"
version = "1.0.0"
description = "Unitree SDK"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "OpenMind", email = "hello@openmind.com"},
]
keywords = ["unitree", "sdk", "go2", "g1"]

dependencies = [
    "Flask==3.1.1",
    "Flask-Cors==6.0.0",
    "Werkzeug==3.1.3",
    "MarkupSafe==3.0.2",
    "click==8.2.1",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.6",
    "docker==7.1.0",
    "requests==2.32.5",
    "urllib3==2.5.0",
    "idna==3.10",
    "certifi==2025.8.3",
    "chardet==5.2.0",
    "websockets==15.0.1",
    "eclipse-zenoh==1.5.0",
    "pyserial==3.5",
    "annotated-types==0.7.0",
    "pydantic==2.11.9",
    "pydantic-core==2.33.2",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "numpy>=1.21.0,<2.0.0",
    "opencv-python>=4.8.0,<4.10.0",
    "apriltag==0.0.16",
]

[project.optional-dependencies]
dev = [
    "flake8>=6.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/OpenMind/unitree-sdk"
Repository = "https://github.com/OpenMind/unitree-sdk"
Documentation = "https://github.com/OpenMind/unitree-sdk/blob/main/README.md"
Issues = "https://github.com/OpenMind/unitree-sdk/issues"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["go2_sdk", "go2_auto_dock", "orchestrator"]
known_third_party = ["rclpy", "ros2", "numpy", "opencv", "flask"]

[tool.flake8]
max-line-length = 88
extend-ignore = [
    "E203",
    "E501",
    "W503",
    "D200",
    "D400"
]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
    ".venv",
    ".tox"
]
per-file-ignores = [
    "__init__.py:F401",  # imported but unused
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

# Ignore missing imports for ROS2 and external packages
[[tool.mypy.overrides]]
module = [
    "rclpy.*",
    "std_msgs.*",
    "geometry_msgs.*",
    "sensor_msgs.*",
    "nav_msgs.*",
    "tf2_ros.*",
    "cv2",
    "numpy",
    "apriltag",
    "docker",
    "websockets",
    "eclipse.zenoh",
    "pyserial",
    "pydantic"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = [
    "go2_sdk/test",
    "go2_auto_dock/test",
    "orchestrator/test"
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.coverage.run]
source = [
    "go2_sdk",
    "go2_auto_dock",
    "orchestrator"
]
omit = [
    "*/test/*",
    "*/tests/*",
    "*/__pycache__/*",
    "*/build/*",
    "*/dist/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:"
]
